Make forest plot with Jasper

make_jasper_forest_plot creates a forest plot with Jasper

make_jasper_forest_plot(
  filestem = "_test",
  headings = NULL,
  rows = NULL,
  cols,
  exponentiate = TRUE,
  col.key = "key",
  col.estimate = "estimate",
  col.stderr = "stderr",
  col.lci = NULL,
  col.uci = NULL,
  col.left = NULL,
  col.right = NULL,
  col.pval = NULL,
  col.keep = NULL,
  xlab = "HR (95% CI)",
  ValueLabelsHeader = "HR (95% CI)",
  forest.xlim = NULL,
  forest.xticks = NULL,
  blankrows = c(1, 1, 0, 0),
  ...
)

Arguments

filestem
String to use for file names. (Default: "_test")

headings
A data frame that contains the headings to be used for the rows of the plot. The data frame must contain columns 'heading1', 'heading2' and 'heading3'. Use NA if a lower level heading is not required.

rows
A character vector. The top level headings (heading1) of rows to be included in the plot.

cols
A list of data frames. These should include columns or point estimates, and standard errors or confidence interval limits. If you specify a headings data frame, then they must also all contain a key column with the same name (which can be specified by col.key).

exponentiate
Exponentiate estimates (and CIs) before plotting. (Default: TRUE)

col.key
Name of column that links the headings provided in headings and the results given in each data frame provided in cols.

If headings data frame is not given, then this column will be used as headings for each row of the plot.

(Default: "key")

col.estimate
Name of column that provides point estimates. (Default: "estimate")

col.stderr
Name of column that provides standard errors. (Default: "stderr")

col.lci
Name of column that provides lower limit of confidence intervals.

col.uci
Name of column that provides upper limit of confidence intervals.

col.left
Names of columns to be printed to the left of the plot.

col.right
Names of columns to be printed to the right of the plot.

col.pval
Name of additional column to be printed to the right of the plot, and formatted as P value.

col.keep
Names of additional columns to be kept in returned data frame.

xlab
Label to appear below the x-axis. (Default: "HR (95% CI)")

ValueLabelsHeader
Label to appear above the text showing estimates and CIs. (Default: "HR (95% CI)")

forest.xlim
A numeric vector. The limits of the x axis.

forest.xticks
A numeric vector. The tick points of the x axis.

blankrows
A numeric vector of length 4 specifying the number of blank rows after a heading1, at the end of a heading1 'section', after a heading2, and at the end of a heading2 'section. (Default: c(1, 1, 0, 0))

...
Other parameters will be passed to the Jasper::ForestFromCSV function.

Details

This function requires the in-house Jasper package.

The function creates a suitable csv, then runs the Jasper::ForestFromCSV to create a forest plot.