Make forest plot with ggplot2

make_forest_plot creates a forest plot with ggplot

make_forest_plot(
  headings = NULL,
  rows = NULL,
  cols,
  exponentiate = TRUE,
  logscale = exponentiate,
  colnames = NULL,
  colheadings = colnames,
  col.key = "key",
  col.estimate = "estimate",
  col.stderr = "stderr",
  col.lci = NULL,
  col.uci = NULL,
  col.left = NULL,
  col.right = NULL,
  col.left.heading = "",
  col.right.heading = "HR (95% CI)",
  col.left.space = 0.02,
  col.right.space = 0.02,
  col.left.hjust = 1,
  col.right.hjust = 0,
  col.heading.space = 0,
  estcolumn = TRUE,
  col.keep = NULL,
  ci.delim = ", ",
  title = "",
  xlab = "HR (95% CI)",
  xlim = NULL,
  xticks = NULL,
  nullval = NULL,
  blankrows = c(1, 1, 0, 0),
  col.diamond = NULL,
  diamond = NULL,
  col.bold = NULL,
  boldheadings = NULL,
  scalepoints = FALSE,
  minse = NULL,
  pointsize = 3,
  shape = NULL,
  colour = NULL,
  cicolour = colour,
  fill = NULL,
  ciunder = NULL,
  addtext = NULL,
  heading.space = 4,
  plot.space = 8,
  base_size = 11,
  base_line_size = base_size/22,
  stroke = base_size/22,
  printplot = TRUE,
  showcode = TRUE
)

Arguments

headings
A data frame that contains the headings to be used for the rows of the plot. The data frame must contain columns 'heading1', 'heading2' and 'heading3'. Use NA if a lower level heading is not required.

rows
A character vector. The top level headings (heading1) of rows to be included in the plot.

cols
A list of data frames. These should include columns or point estimates, and standard errors or confidence interval limits. If you specify a headings data frame, then they must also all contain a key column with the same name (which can be specified by col.key).

exponentiate
Exponentiate estimates (and CIs) before plotting. (Default: TRUE)

logscale
Use log scale on the axis, and add a line at null effect. (Default: exponentiate)

colnames
A character vector. The names to be used for each forest plot. If none provided, then they will be numbered 1, 2, 3 ...

colheadings
Titles to be placed above each forest plot. (Default: colnames)

col.key
Name of column that links the headings provided in headings and the results given in each data frame provided in cols.

If headings data frame is not given, then this column will be used as headings for each row of the plot.

(Default: "key")

col.estimate
Name of column that provides point estimates. (Default: "estimate")

col.stderr
Name of column that provides standard errors. (Default: "stderr")

col.lci
Name of column that provides lower limit of confidence intervals.

col.uci
Name of column that provides upper limit of confidence intervals.

col.left
Names of columns to be printed to the left of the plot.

col.right
Names of columns to be printed to the right of the plot.

col.left.heading
A character vector of titles for col.left columns. (Default: "")

col.right.heading
A character vector of titles for the column of estimates (if estcolumn = TRUE) and col.right columns. (Default: "HR (95% CI)")

col.left.space
A numeric vector. Sizes of the gaps between the plot and col.left columns. As a multiple of the length of the x-axis. (Default: 0)

col.right.space
Size of the gap between the plot and column to the right of the plot. As a multiple of the length of the x-axis. (Default: 0)

col.left.hjust
A numeric vector. The horizontal justification of col.left columns. (Default: 1)

col.right.hjust
A numeric vector. The horizontal justification of col.right columns. (Default: 0)

col.heading.space
Position of the titles given by col.left.heading and col.right.heading. Increase to move them up. (Default: 0)

estcolumn
Include column of estimates and confidence intervals to the right of each plot. (Default: TRUE)

col.keep
Names of additional columns to be kept in returned data frame.

ci.delim
Character string to separate lower and upper limits of confidence interval. (Default: ", ")

title
Title to appear at the top of the plot.

xlab
Label to appear below the x-axis. (Default: "HR (95% CI)")

xlim
A numeric vector. The limits of the x axis.

xticks
A numeric vector. The tick points of the x axis.

nullval
Add a vertical reference line at this value. (If logscale == TRUE then by default it will be added at 1, but use NA not to plot this line.)

blankrows
A numeric vector of length 4 specifying the number of blank rows after a heading1, at the end of a heading1 'section', after a heading2, and at the end of a heading2 'section. (Default: c(1, 1, 0, 0))

col.diamond
Plot estimates and CIs as diamonds. Name of a column of logical values.

diamond
Alternative to col.diamond. A character vectors identify the rows (using the key values) for which the estimate and CI should be plotted using a diamond.

col.bold
Plot text as bold. Name of a column of logical values.

boldheadings
A character vector identifying headings (using key values) which should additionally be bold. (Default: NULL)

scalepoints
Should the points be scaled by inverse of the standard error? (Default: FALSE)

minse
Minimum standard error to use when scaling point size. (Default will use minimum in the data.)

pointsize
The (largest) size of box to use for plotting point estimates. (Default: 3)

shape
Shape of points. An integer, or name of a column of integers. (Default will use shape 22 - squares with fill.)

colour
Colour of points. Name of a colour, or name of a column of colour names. (Default will use black.)

cicolour
Colour of CI lines. Colour of CI lines. Name of a colour, or name of a column of colour names. (Default will use black.)

fill
Fill colour of points. Fill colour of points. Name of a colour, or name of a column of colour names. (Default will use black.)

ciunder
Plot CI lines before points. A logical value, or name of a column of logical values. (Default will plot CI lines after points.)

addtext
A list of data frames. List must be the same length as cols. Data frames should contain a column with the name specified in col.key, and one or more of:

1. a column named 'text' containing character strings

2. columns named 'het_dof', 'het_stat', and 'het_p' containing character strings

3. columns names 'trend_stat' and 'trend_p' containing character strings

The character strings, heterogeneity test, and trend test results will be plotted in the column of estimates and CIs, below the row with the key given in the col.key column.

heading.space
Size of the gap between headings and the first plot. Unit is "lines". (Default: 4)

plot.space
Size of the gap between forest plots. Unit is "lines". (Default: 8)

base_size
base font size, given in pts.

base_line_size
base size for line elements

stroke
Size of outline of shapes. (Default: base_size/22)

printplot
Print the plot. (Default: TRUE)

showcode
Show the ggplot2 code to generate the plot in RStudio 'Viewer' pane. (Default: TRUE)

Value

A list:

plot

the plot

data

a data frame from which the plot is generated

code

ggplot2 code to generate the plot

Details

The function returns the plot, a dataset which is used to create the plot, and the ggplot2 code that creates the plot. In RStudio, the ggplot2 code will be shown in the viewer.